MOD: добавил флаг включенности кеша

+++ models/cache.php
@@ -7,6 +7,8 @@ class Cache {
 	const HOST = '127.0.0.1';
 	const PORT = 11211;
 
+	const CACHE_IS_ENABLED = false;
+
 	/**
 	 * Получить значение кеша
 	 * 
@@ -14,6 +16,10 @@ class Cache {
 	 * @return bool
 	 */
 	public static function getValue($key) {
+	    if (!self::CACHE_IS_ENABLED) {
+	        return false;
+        }
+
 		$memcache = self::_getMemcache();
 		return $memcache->get($key);
 	}
@@ -28,6 +34,10 @@ class Cache {
 	 * @return bool
 	 */
 	public static function setValue($key, $data, $flag = 0, $expire = 604800) {
+        if (!self::CACHE_IS_ENABLED) {
+            return false;
+        }
+
 		$memcache = self::_getMemcache();
 		return $memcache->set($key, $data, $flag, $expire);
 	}
@@ -39,6 +49,10 @@ class Cache {
 	 * @return bool
 	 */
 	public static function deleteValue($key) {
+        if (!self::CACHE_IS_ENABLED) {
+            return false;
+        }
+
 		$memcache = self::_getMemcache();
 		return $memcache->delete($key);
 	}
@@ -55,21 +69,57 @@ class Cache {
 		return $res;
 	}
 
+    public static function getPlayerStatistic() {
+        // ToDo написать логику
+        return false;
+	}
+
+    public static function setPlayerStatistic() {
+        // ToDo написать логику
+        return false;
+    }
+
+    public static function getPlayerInfo() {
+        // ToDo написать логику
+        return false;
+    }
+
+    public static function setPlayerInfo() {
+        // ToDo написать логику
+        return false;
+    }
+
+    public static function getGameInfo() {
+        // ToDo написать логику
+        return false;
+    }
+
+    public static function setGameInfo() {
+        // ToDo написать логику
+        return false;
+    }
+
 	public static function setStatistic($key, $data, $seasonId, $tournamentIds, $flag = 0, $expire = 604800) {
 		$tournamentId = count($tournamentIds) === 1 ? (int) $tournamentIds[0] : 999;
 		$finishData = self::getValue($key);
-		$finishData[$seasonId][$tournamentId] = $data;
+		if ($finishData) {
+            $finishData[$seasonId][$tournamentId] = $data;
 
-		return self::setValue($key, $finishData, $flag, $expire);
+            return self::setValue($key, $finishData, $flag, $expire);
+        }
+		return false;
 	}
 
 	public static function deleteStatistic($key, $seasonId, $tournamentIds) {
 		$tournamentId = count($tournamentIds) === 1 ? (int) $tournamentIds[0] : 999;
 		$data = self::getValue($key);
-		unset($data[$seasonId][$tournamentId]);
-		$res = self::setStatistic($key, $data, $seasonId, $tournamentIds);
+		if ($data) {
+            unset($data[$seasonId][$tournamentId]);
+            $res = self::setStatistic($key, $data, $seasonId, $tournamentIds);
 
-		return $res;
+            return $res;
+        }
+		return false;
 	}
 
 	private static function _getMemcache() {
